refactor : 쿠키 -> shared/model 로 이동 및 setting의 경로 수정
