+++ src/4_Shared/model/cookie.js
